- Update search view - Update state - Update model
