+++ app/src/main/java/com/piashcse/hilt_mvvm_compose_movie/ui/screens/tv_series/tv_series_detail/State.kt
@@ -1,0 +1,14 @@
+package com.piashcse.hilt_mvvm_compose_movie.ui.screens.tv_series.tv_series_detail
+
+import com.piashcse.hilt_mvvm_compose_movie.data.model.TvSeriesItem
+import com.piashcse.hilt_mvvm_compose_movie.data.model.artist.Artist
+import com.piashcse.hilt_mvvm_compose_movie.data.model.tv_series_detail.TvSeriesDetail
+
+data class TvSeriesDetailUiState(
+    val tvSeriesDetail: TvSeriesDetail? = null,
+    val recommendedTvSeries: List<TvSeriesItem> = emptyList(),
+    val tvSeriesCredit: Artist? = null,
+    val isLoading: Boolean = false,
+    val isFavorite: Boolean = false,
+    val errorMessage: String? = null // ✅ Added field
+)
