+++ app/src/main/java/com/piashcse/hilt_mvvm_compose_movie/ui/state/State.kt
@@ -13,28 +13,3 @@ data class UiState(
     val isLoading: Boolean = false,
     val errorMessage: String? = null
 )
-
-data class MovieDetailUiState(
-    val movieDetail: MovieDetail? = null,
-    val recommendedMovies: List<MovieItem> = emptyList(),
-    val movieCredit: Artist? = null,
-    val isLoading: Boolean = false,
-    val isFavorite: Boolean = false,
-    val errorMessage: String? = null // ✅ Added field
-)
-
-data class ArtistDetailUiState(
-    val artistDetail: ArtistDetail? = null,
-    val artistMovies: List<ArtistMovie>? = null,
-    val isLoading: Boolean = false,
-    val errorMessage:  String? = null // ✅ Added error message field
-)
-
-data class TvSeriesDetailUiState(
-    val tvSeriesDetail: TvSeriesDetail? = null,
-    val recommendedTvSeries: List<TvSeriesItem> = emptyList(),
-    val tvSeriesCredit: Artist? = null,
-    val isLoading: Boolean = false,
-    val isFavorite: Boolean = false,
-    val errorMessage: String? = null // ✅ Added field
-)
